requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
langchain>=0.0.350
langchain-community>=0.0.10
pypdf>=3.17.0
python-docx>=0.8.11
openpyxl>=3.1.0
streamlit-extras>=0.3.0 